--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -45,7 +45,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: '3.12'
+          python-version: '3.13'
       - run: python -m pip install -U --user build
       - run: python -m build . --sdist
       - run: python -m pip install -U --user twine
@@ -61,7 +61,7 @@ jobs:
       envs: |
         - windows: py311
         - macos: py312
-        - linux: py310-oldestdeps
+        - linux: py311-oldestdeps
         - linux: py313-devdeps
     secrets:
       CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
@@ -70,7 +70,7 @@ jobs:
     needs: [core]
     uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@v2
     with:
-      default_python: '3.12'
+      default_python: '3.13'
       submodules: false
       pytest: false
       toxdeps: tox-pypi-filter
@@ -90,7 +90,7 @@ jobs:
     needs: [docs]
     uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@main
     with:
-      default_python: '3.12'
+      default_python: '3.13'
       submodules: false
       coverage: codecov
       toxdeps: tox-pypi-filter
@@ -105,7 +105,7 @@ jobs:
           libraries:
             apt:
               - graphviz
-        - linux: py312-online
+        - linux: py313-online
 
   publish:
     # Build wheels on PRs only when labelled. Releases will only be published if tagged ^v.*
@@ -119,7 +119,7 @@ jobs:
     needs: [test, docs]
     uses: OpenAstronomy/github-actions-workflows/.github/workflows/publish_pure_python.yml@v2
     with:
-      python-version: '3.12'
+      python-version: '3.13'
       test_extras: 'all,tests'
       test_command: 'pytest -p no:warnings --doctest-rst --pyargs aiapy'
       submodules: false