--- a/aiapy/calibrate/prep.py
+++ b/aiapy/calibrate/prep.py
@@ -178,9 +178,11 @@ def degradation(
     All calibration terms are taken from the ``aia.response`` series in JSOC
     or read from the table input by the user.
 
-    .. note:: This function is adapted directly from the
-              `aia_bp_corrections.pro <https://sohoftp.nascom.nasa.gov/solarsoft/sdo/aia/idl/response/aia_bp_corrections.pro>`__
-              routine in SolarSoft.
+    .. note::
+
+        This function is adapted directly from the
+        `aia_bp_corrections.pro <https://sohoftp.nascom.nasa.gov/solarsoft/sdo/aia/idl/response/aia_bp_corrections.pro>`__
+        routine in SolarSoft.
 
     Parameters
     ----------