--- a/aiapy/psf/psf.py
+++ b/aiapy/psf/psf.py
@@ -11,11 +11,11 @@ from sunpy import log
 from aiapy.util.decorators import validate_channel
 
 try:
-    import cupy
-
-    HAS_CUPY = True
+    from jax import jit as jax_jit
+    import jax.numpy as jnp
 except ImportError:
-    HAS_CUPY = False
+    jnp = np
+    def jax_jit(func): return func
 
 __all__ = ["_psf", "filter_mesh_parameters", "psf"]
 
@@ -174,7 +174,7 @@ def filter_mesh_parameters(*, use_preflightcore=False):
 
 @u.quantity_input
 @validate_channel("channel", valid_channels=[94, 131, 171, 193, 211, 304, 335] * u.angstrom)
-def psf(channel: u.angstrom, *, use_preflightcore=False, diffraction_orders=None, use_gpu=True):
+def psf(channel: u.angstrom, *, use_preflightcore=False, diffraction_orders=None):
     r"""
     Calculate the composite PSF for a given channel, including diffraction and
     core effects.
@@ -182,11 +182,8 @@ def psf(channel: u.angstrom, *, use_preflightcore=False, diffraction_orders=None
     .. note:: This function has been adapted from
               `aia_calc_psf.pro <https://sohoftp.nascom.nasa.gov/solarsoft/sdo/aia/idl/psf/PRO/aia_calc_psf.pro>`_.
 
-    .. note:: If the `~cupy` package is installed
-              and your machine has an NVIDIA GPU, the PSF calculation will
-              automatically be accelerated with CUDA. This can lead to
-              several orders of magnitude in performance increase compared to
-              pure `numpy` on a CPU.
+    .. note:: If the `~jax` package is installed it will be used to accelerate the computation.
+              `~jax` can use CPUs or GPUs, see their documentation for instructions.
 
     The point spread function (PSF) can be modeled as a 2D Gaussian function
     of the radial distance :math:`r` from the center,
@@ -249,9 +246,6 @@ def psf(channel: u.angstrom, *, use_preflightcore=False, diffraction_orders=None
     diffraction_orders : array-like, optional
         The diffraction orders to sum over. If None, the full
         range from -100 to +100 in steps of 1 will be used.
-    use_gpu : `bool`, optional
-        If True and `~cupy` is installed, do PSF deconvolution on the GPU
-        with `~cupy`.
 
     Returns
     -------
@@ -275,56 +269,48 @@ def psf(channel: u.angstrom, *, use_preflightcore=False, diffraction_orders=None
     angles_focal_plane = u.Quantity([45.0, -45.0], "deg")
     if diffraction_orders is None:
         diffraction_orders = np.arange(-100, 101, 1)
-    psf_entrance = _psf(meshinfo, angles_entrance, diffraction_orders, use_gpu=use_gpu)
+    psf_entrance = _psf(meshinfo, angles_entrance, diffraction_orders)
     psf_focal_plane = _psf(
         meshinfo,
         angles_focal_plane,
         diffraction_orders,
         focal_plane=True,
-        use_gpu=use_gpu,
     )
     # Composite PSF
     psf = abs(np.fft.fft2(np.fft.fft2(psf_focal_plane) * np.fft.fft2(psf_entrance)))
     # Center PSF in the middle of the image
     psf = np.roll(np.roll(psf, psf.shape[1] // 2, axis=1), psf.shape[0] // 2, axis=0)
     # Normalize by total number of pixels
-    psf = psf / (psf.shape[0] * psf.shape[1])
-    # If using cupy, cast back to a normal numpy array
-    if HAS_CUPY and use_gpu:
-        psf = cupy.asnumpy(psf)
-    return psf
+    return psf / (psf.shape[0] * psf.shape[1])
 
 
-def _psf(meshinfo, angles, diffraction_orders, *, focal_plane=False, use_gpu=True):
-    psf = np.zeros((4096, 4096), dtype=float)
-    if use_gpu and not HAS_CUPY:
-        log.info("cupy not installed or working, falling back to CPU")
-    # If cupy is available, cast to a cupy array
-    if HAS_CUPY and use_gpu:
-        psf = cupy.array(psf)
+@jax_jit
+def _psf_inner(x, y, order, mesh_ratio, spacing_x, spacing_y, nx, ny, width_x, width_y):
+    intensity = jnp.sinc(order / mesh_ratio) ** 2  # I_0
+    for dx, dy in zip(spacing_x, spacing_y, strict=True):
+        x_centered = x - (0.5 * nx + dx * order + 0.5)
+        y_centered = y - (0.5 * ny + dy * order + 0.5)
+        return jnp.exp(-width_x * x_centered * x_centered - width_y * y_centered * y_centered) * intensity
+    return None
+
+
+def _psf(meshinfo, angles, diffraction_orders, *, focal_plane=False):
+    psf = jnp.zeros((4096, 4096), dtype=float)
     nx, ny = psf.shape
     width_x = meshinfo["width"].value
     width_y = meshinfo["width"].value
     # x and y position grids
-    x = np.outer(np.ones(ny), np.arange(nx) + 0.5)
-    y = np.outer(np.arange(ny) + 0.5, np.ones(nx))
-    if HAS_CUPY and use_gpu:
-        x = cupy.array(x)
-        y = cupy.array(y)
+    x = jnp.outer(jnp.ones(ny), jnp.arange(nx) + 0.5)
+    y = jnp.outer(jnp.arange(ny) + 0.5, jnp.ones(nx))
     area_not_mesh = 0.82  # fractional area not covered by the mesh
     spacing = meshinfo["spacing_fp"] if focal_plane else meshinfo["spacing_e"]
     mesh_ratio = (meshinfo["mesh_pitch"] / meshinfo["mesh_width"]).decompose().value
-    spacing_x = spacing * np.cos(angles)
-    spacing_y = spacing * np.sin(angles)
+    spacing_x = spacing.value * jnp.cos(angles)
+    spacing_y = spacing.value * jnp.sin(angles)
     for order in diffraction_orders:
         if order == 0:
             continue
-        intensity = np.sinc(order / mesh_ratio) ** 2  # I_0
-        for dx, dy in zip(spacing_x.value, spacing_y.value, strict=True):
-            x_centered = x - (0.5 * nx + dx * order + 0.5)
-            y_centered = y - (0.5 * ny + dy * order + 0.5)
-            # NOTE: this step is the bottleneck and is VERY slow on a CPU
-            psf += np.exp(-width_x * x_centered * x_centered - width_y * y_centered * y_centered) * intensity
+        psf += _psf_inner(x, y, order, mesh_ratio, spacing_x, spacing_y, nx, ny, width_x, width_y)
     # Contribution from core
-    psf_core = np.exp(-width_x * (x - 0.5 * nx - 0.5) ** 2 - width_y * (y - 0.5 * ny - 0.5) ** 2)
+    psf_core = jnp.exp(-width_x * (x - 0.5 * nx - 0.5) ** 2 - width_y * (y - 0.5 * ny - 0.5) ** 2)
     return (1 - area_not_mesh) * psf / psf.sum() + area_not_mesh * psf_core / psf_core.sum()