--- a/pyproject.toml
+++ b/pyproject.toml
@@ -9,7 +9,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "aiapy"
 description = "Python library for AIA data analysis."
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 readme = { file = "README.rst", content-type = "text/x-rst" }
 license = { file = "licenses/LICENSE.rst" }
 authors = [
@@ -37,7 +37,6 @@ classifiers = [
   "Operating System :: OS Independent",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: 3.13",
@@ -46,6 +45,9 @@ classifiers = [
 
 [project.optional-dependencies]
 all = ["aiapy"]
+performance = ["jax"]
+performance_nvidia = ["jax[cuda13]"]
+performance_mac = ["jax-metal"]
 tests = [
   "pytest",
   "pytest-doctestplus",