--- a/tox.ini
+++ b/tox.ini
@@ -6,7 +6,6 @@ envlist =
     py{310,311,312,313}
     py313-devdeps
     py310-oldestdeps
-    cupy
     codestyle
     build_docs{,-gallery}
 
@@ -32,9 +31,6 @@ set_env =
     MPLBACKEND = agg
     devdeps: PIP_EXTRA_INDEX_URL = https://pypi.anaconda.org/astropy/simple https://pypi.anaconda.org/scientific-python-nightly-wheels/simple
 deps =
-    # Pin due to parsing long times in astropy with this version of numpy
-    numpy<2.3.0
-    cupy: cupy-cuda12x
     # For packages which publish nightly wheels this will pull the latest nightly
     devdeps: astropy>=0.0.dev0
     devdeps: sunpy>=0.0.dev0