--- a/tox.ini
+++ b/tox.ini
@@ -3,9 +3,9 @@ min_version = 4.0
 requires =
     tox-pypi-filter>=0.14
 envlist =
-    py{310,311,312,313}
+    py{311,312,313}
     py313-devdeps
-    py310-oldestdeps
+    py311-oldestdeps
     codestyle
     build_docs{,-gallery}