This makes a massive difference in CPU compute time.

Before:
```
6min 34s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

After:
```
21.2 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

They don't give *exactly* the same numerical answers, but reasonably close:
```
Not equal to tolerance rtol=1e-07, atol=0

Mismatched elements: 16777216 / 16777216 (100%)
Max absolute difference among violations: 0.01832044
Max relative difference among violations: 1.55556522e+18
 ACTUAL: array([[8.981260e-12, 1.183608e-13, 3.165574e-12, ..., 6.444088e-13,
        3.165583e-12, 1.182429e-13],
       [7.738048e-12, 8.832175e-12, 1.016188e-11, ..., 1.666469e-11,...
 DESIRED: array([[2.574981e-19, 1.831484e-20, 1.138003e-20, ..., 2.011767e-21,
        1.752931e-20, 3.083227e-20],
       [2.744908e-19, 3.670137e-20, 2.307091e-21, ..., 3.173223e-20,...
```

## Summary by Sourcery

Switch PSF computation from Cupy to JAX for significant performance improvements, refactoring the core functions to use jax.jit and jax.numpy and removing legacy Cupy support.

Enhancements:
- Replace Cupy-based GPU acceleration with JAX-based acceleration using jax.jit and jax.numpy throughout the PSF computation
- Remove the use_gpu parameter and associated Cupy fallback logic from psf and _psf functions
- Wrap the diffraction computation in a jax_jit-decorated inner function to enable JIT compilation
- Update import logic to fallback to NumPy when JAX is unavailable

Documentation:
- Update documentation to reference JAX acceleration and remove Cupy mentions