@nabobalis 2025-09-15
When presented with a hog, always go full hog.