@nabobalis 2025-09-15
I will have to put this down for today but I want to come back to this and go further.