@nabobalis 2025-09-15
Yeah, I will do a plot before and after.